+++ store/bbolt_test.go
@@ -1,4 +1,4 @@
-package scheduler
+package store
 
 import (
 	"sync"
