+++ store/store.go
@@ -1,0 +1,7 @@
+package store
+
+type Store interface {
+	Get([]byte) ([]byte, error)
+	Put([]byte, []byte) error
+	Delete([]byte) error
+}
